+++ rector.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 use Rector\Core\Configuration\Option;
 use Rector\Core\ValueObject\PhpVersion;
+use Rector\Php71\Rector\FuncCall\CountOnNullRector;
 use Rector\Set\ValueObject\SetList;
 use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
 
@@ -27,6 +28,7 @@ return static function (ContainerConfigurator $containerConfigurator): void {
     $parameters->set(Option::SKIP, [
         __DIR__ . '/src/Serializer',
         __DIR__ . '/vendor',
+        CountOnNullRector::class
     ]);
 
     $parameters->set(Option::SETS, [