+++ src/Exception/RuntimeException.php
@@ -4,9 +4,11 @@ declare(strict_types=1);
 
 namespace SchedulerBundle\Exception;
 
+use RuntimeException as InternalRuntimeException;
+
 /**
  * @author Guillaume Loulier <contact@guillaumeloulier.fr>
  */
-class RuntimeException extends \RuntimeException implements ExceptionInterface
+class RuntimeException extends InternalRuntimeException implements ExceptionInterface
 {
 }