+++ src/Middleware/AbstractMiddlewareStack.php
@@ -6,7 +6,7 @@ namespace SchedulerBundle\Middleware;
 
 use Closure;
 use function array_filter;
-use function array_merge;
+use function array_replace;
 use function uasort;
 
 /**
@@ -14,16 +14,22 @@ use function uasort;
  */
 abstract class AbstractMiddlewareStack implements MiddlewareStackInterface
 {
+    /**
+     * @var iterable|PostExecutionMiddlewareInterface[]|PostSchedulingMiddlewareInterface[]|PreExecutionMiddlewareInterface[]|PreSchedulingMiddlewareInterface[]|OrderedMiddlewareInterface[]
+     */
     protected iterable $stack;
 
     /**
-     * @param iterable|PreExecutionMiddlewareInterface[]|PostExecutionMiddlewareInterface[]|PreSchedulingMiddlewareInterface[]|PostSchedulingMiddlewareInterface[] $stack
+     * @param iterable|PreExecutionMiddlewareInterface[]|PostExecutionMiddlewareInterface[]|PreSchedulingMiddlewareInterface[]|PostSchedulingMiddlewareInterface[]|OrderedMiddlewareInterface[] $stack
      */
     public function __construct(iterable $stack = [])
     {
         $this->stack = $stack;
     }
 
+    /**
+     * @return PreSchedulingMiddlewareInterface[]|OrderedMiddlewareInterface[]
+     */
     protected function getPreSchedulingMiddleware(): array
     {
         $list = array_filter($this->stack, fn ($middleware): bool => $middleware instanceof PreSchedulingMiddlewareInterface);
@@ -31,6 +37,9 @@ abstract class AbstractMiddlewareStack implements MiddlewareStackInterface
         return $this->orderMiddleware($list);
     }
 
+    /**
+     * @return PostSchedulingMiddlewareInterface[]|OrderedMiddlewareInterface[]
+     */
     protected function getPostSchedulingMiddleware(): array
     {
         $list = array_filter($this->stack, fn ($middleware): bool => $middleware instanceof PostSchedulingMiddlewareInterface);
@@ -38,6 +47,9 @@ abstract class AbstractMiddlewareStack implements MiddlewareStackInterface
         return $this->orderMiddleware($list);
     }
 
+    /**
+     * @return PreExecutionMiddlewareInterface[]|OrderedMiddlewareInterface[]
+     */
     protected function getPreExecutionMiddleware(): array
     {
         $list = array_filter($this->stack, fn ($middleware): bool => $middleware instanceof PreExecutionMiddlewareInterface);
@@ -45,6 +57,9 @@ abstract class AbstractMiddlewareStack implements MiddlewareStackInterface
         return $this->orderMiddleware($list);
     }
 
+    /**
+     * @return PostExecutionMiddlewareInterface[]|OrderedMiddlewareInterface[]
+     */
     protected function getPostExecutionMiddleware(): array
     {
         $list = array_filter($this->stack, fn ($middleware): bool => $middleware instanceof PostExecutionMiddlewareInterface);
@@ -52,6 +67,10 @@ abstract class AbstractMiddlewareStack implements MiddlewareStackInterface
         return $this->orderMiddleware($list);
     }
 
+    /**
+     * @param PostExecutionMiddlewareInterface[]|PostSchedulingMiddlewareInterface[]|PreExecutionMiddlewareInterface[]|PreSchedulingMiddlewareInterface[]|OrderedMiddlewareInterface[] $middlewareList
+     * @param Closure                                                                                                                                                                  $func
+     */
     protected function runMiddleware(array $middlewareList, Closure $func): void
     {
         array_walk($middlewareList, $func);