+++ tests/Worker/WorkerTest.php
@@ -13,8 +13,10 @@ use SchedulerBundle\Middleware\RateLimiterMiddleware;
 use SchedulerBundle\Middleware\TaskCallbackMiddleware;
 use SchedulerBundle\Middleware\WorkerMiddlewareStack;
 use SchedulerBundle\Task\FailedTask;
+use SchedulerBundle\Task\TaskListInterface;
 use SchedulerBundle\TaskBag\NotificationTaskBag;
 use Symfony\Component\EventDispatcher\EventDispatcher;
+use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 use Symfony\Component\Lock\BlockingStoreInterface;
 use SchedulerBundle\EventListener\StopWorkerOnTaskLimitSubscriber;
 use SchedulerBundle\Exception\UndefinedRunnerException;
@@ -31,7 +33,6 @@ use Symfony\Component\Notifier\NotifierInterface;
 use Symfony\Component\Notifier\Recipient\Recipient;
 use Symfony\Component\RateLimiter\RateLimiterFactory;
 use Symfony\Component\RateLimiter\Storage\InMemoryStorage;
-use Symfony\Contracts\EventDispatcher\EventDispatcherInterface;
 
 /**
  * @author Guillaume Loulier <contact@guillaumeloulier.fr>
@@ -86,7 +87,7 @@ final class WorkerTest extends TestCase
         $scheduler->expects(self::once())->method('getDueTasks')->willReturn(new TaskList([$task]));
 
         $runner = $this->createMock(RunnerInterface::class);
-        $runner->expects(self::once())->method('support')->willReturn(false);
+        $runner->expects(self::once())->method('support')->with(self::equalTo($task))->willReturn(false);
         $runner->expects(self::never())->method('run');
 
         $secondRunner = $this->createMock(RunnerInterface::class);
@@ -136,10 +137,6 @@ final class WorkerTest extends TestCase
             ]
         );
 
-        $tracker = $this->createMock(TaskExecutionTrackerInterface::class);
-        $tracker->expects(self::once())->method('startTracking');
-        $tracker->expects(self::once())->method('endTracking');
-
         $task = $this->createMock(TaskInterface::class);
         $task->expects(self::exactly(3))->method('getName')->willReturn('foo');
         $task->expects(self::once())->method('getExpression')->willReturn('* * * * *');
@@ -153,6 +150,10 @@ final class WorkerTest extends TestCase
         $secondTask->expects(self::once())->method('setExecutionStartTime');
         $secondTask->expects(self::once())->method('setExecutionEndTime');
         $secondTask->expects(self::once())->method('setLastExecution');
+
+        $tracker = $this->createMock(TaskExecutionTrackerInterface::class);
+        $tracker->expects(self::once())->method('startTracking')->with(self::equalTo($secondTask));
+        $tracker->expects(self::once())->method('endTracking')->with(self::equalTo($secondTask));
 
         $runner = $this->createMock(RunnerInterface::class);
         $runner->expects(self::once())->method('support')->with($secondTask)->willReturn(true);
@@ -215,10 +216,6 @@ final class WorkerTest extends TestCase
         $logger = $this->createMock(LoggerInterface::class);
         $logger->expects(self::never())->method('info');
 
-        $tracker = $this->createMock(TaskExecutionTrackerInterface::class);
-        $tracker->expects(self::never())->method('startTracking');
-        $tracker->expects(self::never())->method('endTracking');
-
         $task = $this->createMock(TaskInterface::class);
         $task->expects(self::exactly(4))->method('getName')->willReturn('foo');
         $task->expects(self::exactly(2))->method('getState')->willReturn(TaskInterface::ENABLED);
@@ -229,8 +226,12 @@ final class WorkerTest extends TestCase
         $task->expects(self::never())->method('setExecutionEndTime');
         $task->expects(self::never())->method('setLastExecution');
 
+        $tracker = $this->createMock(TaskExecutionTrackerInterface::class);
+        $tracker->expects(self::never())->method('startTracking')->with(self::equalTo($task));
+        $tracker->expects(self::never())->method('endTracking')->with(self::equalTo($task));
+
         $runner = $this->createMock(RunnerInterface::class);
-        $runner->expects(self::once())->method('support')->with($task)->willReturn(true);
+        $runner->expects(self::once())->method('support')->with(self::equalTo($task))->willReturn(true);
         $runner->expects(self::never())->method('run');
 
         $scheduler = $this->createMock(SchedulerInterface::class);
@@ -254,10 +255,6 @@ final class WorkerTest extends TestCase
         $logger = $this->createMock(LoggerInterface::class);
         $logger->expects(self::never())->method('info');
 
-        $tracker = $this->createMock(TaskExecutionTrackerInterface::class);
-        $tracker->expects(self::once())->method('startTracking');
-        $tracker->expects(self::once())->method('endTracking');
-
         $task = $this->createMock(TaskInterface::class);
         $task->expects(self::exactly(4))->method('getName')->willReturn('foo');
         $task->expects(self::exactly(2))->method('getState')->willReturn(TaskInterface::ENABLED);
@@ -278,6 +275,10 @@ final class WorkerTest extends TestCase
         $validTask->expects(self::once())->method('setExecutionStartTime');
         $validTask->expects(self::once())->method('setExecutionEndTime');
         $validTask->expects(self::once())->method('setLastExecution');
+
+        $tracker = $this->createMock(TaskExecutionTrackerInterface::class);
+        $tracker->expects(self::once())->method('startTracking')->with(self::equalTo($validTask));
+        $tracker->expects(self::once())->method('endTracking')->with(self::equalTo($validTask));
 
         $runner = $this->createMock(RunnerInterface::class);
         $runner->expects(self::exactly(2))->method('support')->withConsecutive([$task], [$validTask])->willReturn(true);
@@ -304,10 +305,6 @@ final class WorkerTest extends TestCase
         $logger = $this->createMock(LoggerInterface::class);
         $logger->expects(self::never())->method('info');
 
-        $tracker = $this->createMock(TaskExecutionTrackerInterface::class);
-        $tracker->expects(self::once())->method('startTracking');
-        $tracker->expects(self::once())->method('endTracking');
-
         $task = $this->createMock(TaskInterface::class);
         $task->expects(self::exactly(2))->method('getName')->willReturn('foo');
         $task->expects(self::exactly(2))->method('getState')->willReturn(TaskInterface::ENABLED);
@@ -317,6 +314,10 @@ final class WorkerTest extends TestCase
         $task->expects(self::once())->method('setExecutionStartTime');
         $task->expects(self::once())->method('setExecutionEndTime');
         $task->expects(self::once())->method('setLastExecution');
+
+        $tracker = $this->createMock(TaskExecutionTrackerInterface::class);
+        $tracker->expects(self::once())->method('startTracking')->with(self::equalTo($task));
+        $tracker->expects(self::once())->method('endTracking')->with(self::equalTo($task));
 
         $runner = $this->createMock(RunnerInterface::class);
         $runner->expects(self::once())->method('support')->with($task)->willReturn(true);
@@ -535,14 +536,14 @@ final class WorkerTest extends TestCase
         $runner->expects(self::once())->method('support')->willReturn(true);
         $runner->expects(self::once())->method('run')->willThrowException(new RuntimeException('Random error occurred'));
 
-        $tracker = $this->createMock(TaskExecutionTrackerInterface::class);
-        $tracker->expects(self::once())->method('startTracking');
-        $tracker->expects(self::never())->method('endTracking');
-
         $logger = $this->createMock(LoggerInterface::class);
 
         $task = $this->createMock(TaskInterface::class);
         $task->method('getName')->willReturn('failed');
+
+        $tracker = $this->createMock(TaskExecutionTrackerInterface::class);
+        $tracker->expects(self::once())->method('startTracking')->with(self::equalTo($task));
+        $tracker->expects(self::never())->method('endTracking');
 
         $scheduler = $this->createMock(SchedulerInterface::class);
         $scheduler->expects(self::never())->method('getTimezone');
@@ -646,7 +647,7 @@ final class WorkerTest extends TestCase
         $worker = new Worker($scheduler, [$runner], $tracker, new WorkerMiddlewareStack([
             new TaskCallbackMiddleware(),
             new NotifierMiddleware($notifier),
-        ]), $eventDispatcher, $logger, null);
+        ]), $eventDispatcher, $logger);
         $worker->execute();
 
         self::assertSame($task, $worker->getLastExecutedTask());
@@ -689,7 +690,7 @@ final class WorkerTest extends TestCase
         $worker = new Worker($scheduler, [$runner], $tracker, new WorkerMiddlewareStack([
             new TaskCallbackMiddleware(),
             new NotifierMiddleware($notifier),
-        ]), $eventDispatcher, $logger, null);
+        ]), $eventDispatcher, $logger);
         $worker->execute();
 
         self::assertSame($task, $worker->getLastExecutedTask());
@@ -735,7 +736,7 @@ final class WorkerTest extends TestCase
         $worker = new Worker($scheduler, [$runner], $tracker, new WorkerMiddlewareStack([
             new TaskCallbackMiddleware(),
             new NotifierMiddleware($notifier),
-        ]), $eventDispatcher, $logger, null);
+        ]), $eventDispatcher, $logger);
         $worker->execute();
 
         self::assertSame($task, $worker->getLastExecutedTask());
@@ -873,5 +874,75 @@ final class WorkerTest extends TestCase
         self::assertSame($task, $worker->getLastExecutedTask());
         self::assertSame($task, $failedTask->getTask());
         self::assertSame('Rate Limit Exceeded', $failedTask->getReason());
+    }
+
+    public function testWorkerCanStopWhenTaskAreConsumedAndWithoutDaemonEnabled(): void
+    {
+        $logger = $this->createMock(LoggerInterface::class);
+        $logger->expects(self::never())->method('info');
+
+        $task = $this->createMock(TaskInterface::class);
+        $task->expects(self::once())->method('getName')->willReturn('foo');
+        $task->expects(self::exactly(2))->method('getState')->willReturn(TaskInterface::ENABLED);
+        $task->expects(self::once())->method('isSingleRun')->willReturn(false);
+        $task->expects(self::once())->method('setArrivalTime');
+        $task->expects(self::once())->method('setExecutionStartTime');
+        $task->expects(self::once())->method('setExecutionEndTime');
+        $task->expects(self::once())->method('setLastExecution');
+
+        $tracker = $this->createMock(TaskExecutionTrackerInterface::class);
+        $tracker->expects(self::once())->method('startTracking')->with(self::equalTo($task));
+        $tracker->expects(self::once())->method('endTracking')->with(self::equalTo($task));
+
+        $runner = $this->createMock(RunnerInterface::class);
+        $runner->expects(self::once())->method('support')->with($task)->willReturn(true);
+        $runner->expects(self::once())->method('run')->with($task)->willReturn(new Output($task, null));
+
+        $scheduler = $this->createMock(SchedulerInterface::class);
+
+        $eventDispatcher = $this->createMock(EventDispatcherInterface::class);
+        $eventDispatcher->expects(self::exactly(7))->method('dispatch');
+
+        $worker = new Worker($scheduler, [$runner], $tracker, new WorkerMiddlewareStack(), $eventDispatcher, $logger);
+        $worker->execute([], $task);
+
+        self::assertInstanceOf(TaskListInterface::class, $worker->getFailedTasks());
+        self::assertEmpty($worker->getFailedTasks());
+        self::assertSame($task, $worker->getLastExecutedTask());
+    }
+
+    public function testWorkerCanStopWhenTaskAreConsumedWithError(): void
+    {
+        $logger = $this->createMock(LoggerInterface::class);
+        $logger->expects(self::never())->method('info');
+
+        $task = $this->createMock(TaskInterface::class);
+        $task->expects(self::exactly(2))->method('getName')->willReturn('foo');
+        $task->expects(self::exactly(2))->method('getState')->willReturn(TaskInterface::ENABLED);
+        $task->expects(self::once())->method('isSingleRun')->willReturn(false);
+        $task->expects(self::once())->method('setArrivalTime');
+        $task->expects(self::once())->method('setExecutionStartTime');
+        $task->expects(self::never())->method('setExecutionEndTime');
+        $task->expects(self::never())->method('setLastExecution');
+
+        $tracker = $this->createMock(TaskExecutionTrackerInterface::class);
+        $tracker->expects(self::once())->method('startTracking')->with(self::equalTo($task));
+        $tracker->expects(self::never())->method('endTracking');
+
+        $runner = $this->createMock(RunnerInterface::class);
+        $runner->expects(self::once())->method('support')->with($task)->willReturn(true);
+        $runner->expects(self::once())->method('run')->with($task)->willThrowException(new RuntimeException('An error occurred'));
+
+        $scheduler = $this->createMock(SchedulerInterface::class);
+
+        $eventDispatcher = $this->createMock(EventDispatcherInterface::class);
+        $eventDispatcher->expects(self::exactly(7))->method('dispatch');
+
+        $worker = new Worker($scheduler, [$runner], $tracker, new WorkerMiddlewareStack(), $eventDispatcher, $logger);
+        $worker->execute([], $task);
+
+        self::assertInstanceOf(TaskListInterface::class, $worker->getFailedTasks());
+        self::assertNotEmpty($worker->getFailedTasks());
+        self::assertSame($task, $worker->getLastExecutedTask());
     }
 }